+++ app.js
@@ -1,0 +1,18 @@
+const express = require('express')
+const app = express()
+const cors = require('cors')
+const router = require('./src/routes/routes')
+const port = 3000
+app.use(cors())
+app.use(express.json())
+app.use(router)
+app.set('view engine','ejs')
+
+app.get('/', (req, res) =>{
+    const number = Math.random()
+    response.render(path.join(__dirname, 'src/views/index.ejs'),{variavel: number})
+})
+
+app.listen(port,()=>{
+    console.log(`servidor rodando na porta http://localhost:${port}`)
+})
